+++ CHANGELOG.md
@@ -24,7 +24,7 @@ This is a small bugfix release for some issues with the new experimental Circuit
 - **[FIR]** Better ensure `enableCircuitCodegen` and `generateContributionProviders` work together when both enabled.
 - **[IR]** Fix default parameter expressions not being copied when `generateContributionProviders` is enabled. This specifically affected scoped or private bindings.
 - **[IR]** Fix qualifier annotations not being copied when `generateContributionProviders` is enabled.
-- **[IR]** Don't link expect/actual declarations if the callee is a synthetic declaration. This avoid some non-obvious IC failures with `generateContributionProviders`.
+- **[IR]** Don't link expect/actual declarations if the callee is a synthetic declaration. This avoids some non-obvious IC failures with `generateContributionProviders`.
 
 ### Contributors